pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepshield"
version = "0.1.0"
description = "DeepShield OpenEnv security benchmark environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "gradio>=5.0.0",
    "huggingface_hub>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
